feat: support custom hook for checking access
Pass product and issue to the hook. Plugin user is responsible for
registering hook and provide a logic for determining logged in user
access. As a default, access is denied.

Default shortcode template doesn't check access, every issue has a link
to the redirect page. Template can be customized for more advanced logic.